split CodeGenerator helpers into separate units
